Fix: Implement proper case-insensitive and date filtering logic in PandasDataAccessor
- Updated filtering logic to handle case-insensitive comparisons and string operations.
- Fixed issue with date filtering by correctly handling datetime conversion.
- Resolved test failures for contains and equals operations in case-insensitive string filtering.
- All test cases now pass successfully.
