+++ taipy/gui/data/pandas_data_accessor.py
@@ -269,25 +269,46 @@ class _PandasDataAccessor(_DataAccessor):
                 action = fd.get("action")
                 match_case = fd.get("matchCase", False)
 
-                if isinstance(val, str):
+                # Handle date columns separately from query
+                if self.__is_date_column(t.cast(pd.DataFrame, df), col):
+                    val = datetime.fromisoformat(val[:-1])  # Convert to datetime
+                    if action == ">":
+                        df = df[df[col] > val]
+                    elif action == "<":
+                        df = df[df[col] < val]
+                    elif action == "==":
+                        df = df[df[col] == val]
+                    is_copied = True
+                    continue  # Skip query for dates since it's already applied
+
+                elif isinstance(val, str):
+                    # Handle string operations
                     if action == "contains":
                         if not match_case:
-                            val = val.lower()  # Convert value to lowercase for case-insensitive filtering
+                            col = f"`{col}`.str.lower()"
+                            val = val.lower()
                         vars.append(val)
-                        # Apply case-sensitive or case-insensitive string filtering
                         right = f".str.contains(@vars[{len(vars) - 1}], case={match_case})"
                     else:
+                        if not match_case:
+                            col = f"`{col}`.str.lower()"
+                            val = val.lower()
                         vars.append(val)
-                    val = f"@vars[{len(vars) - 1}]" if isinstance(val, (str, datetime)) else val
+                        right = f" {action} @vars[{len(vars) - 1}]"
                 else:
+                    # Handle numeric or other types
                     vars.append(val)
-                right = f" {action} {val}" if action != "contains" else right  # type: ignore
+                    right = f" {action} @vars[{len(vars) - 1}]"
+
                 if query:
                     query += " and "
-                query += f"`{col}`{right}"
+                query += f"{col}{right}"
+
+            # Apply string/numeric filters using df.query()
             try:
-                df = df.query(query)
-                is_copied = True
+                if query:
+                    df = df.query(query)
+                    is_copied = True
             except Exception as e:
                 _warn(f"Dataframe filtering: invalid query '{query}' on {df.head()}", e)
