+++ tests/gui/data/test_pandas_data_accessor.py
@@ -16,12 +16,35 @@ from importlib import util
 
 import pandas
 from flask import g
+import pandas as pd
 
 from taipy.gui import Gui
 from taipy.gui.data.data_format import _DataFormat
 from taipy.gui.data.decimator import ScatterDecimator
 from taipy.gui.data.pandas_data_accessor import _PandasDataAccessor
 
+import pytest
+from unittest.mock import Mock
+
+# Define a mock to simulate _DataFormat behavior with a 'value' attribute
+class MockDataFormat:
+    LIST = Mock(value="list")
+    CSV = Mock(value="csv")
+
+@pytest.fixture
+def pandas_accessor():
+    gui = Mock()  
+    return _PandasDataAccessor(gui=gui)
+
+@pytest.fixture
+def sample_df():
+    data = {
+        "StringCol": ["Apple", "Banana", "Cherry", "apple"],
+        "NumberCol": [10, 20, 30, 40],
+        "BoolCol": [True, False, True, False],
+        "DateCol": pd.to_datetime(["2020-01-01", "2021-06-15", "2022-08-22", "2023-03-05"])
+    }
+    return pd.DataFrame(data)
 
 def test_simple_data(gui: Gui, helpers, small_dataframe):
     accessor = _PandasDataAccessor(gui)
@@ -255,6 +278,58 @@ def test_filter_by_date(gui: Gui, helpers, small_dataframe):
     value = accessor.get_data("x", pd, query, _DataFormat.JSON)
     assert len(value["value"]["data"]) == 1
 
+def test_contains_case_sensitive(pandas_accessor, sample_df):
+    payload = {
+        "filters": [{"col": "StringCol", "value": "Apple", "action": "contains", "matchCase": True}]
+    }
+    result = pandas_accessor.get_data("test_var", sample_df, payload, MockDataFormat.LIST)
+    filtered_data = pd.DataFrame(result['value']['data'])
+    
+    assert len(filtered_data) == 1
+    assert filtered_data.iloc[0]['StringCol'] == 'Apple'
+
+def test_contains_case_insensitive(pandas_accessor, sample_df):
+    payload = {
+        "filters": [{"col": "StringCol", "value": "apple", "action": "contains", "matchCase": False}]
+    }
+    result = pandas_accessor.get_data("test_var", sample_df, payload, MockDataFormat.LIST)
+    filtered_data = pd.DataFrame(result['value']['data'])
+    
+    assert len(filtered_data) == 2
+    assert 'Apple' in filtered_data['StringCol'].values
+    assert 'apple' in filtered_data['StringCol'].values
+
+def test_equals_case_sensitive(pandas_accessor, sample_df):
+    payload = {
+        "filters": [{"col": "StringCol", "value": "Apple", "action": "==", "matchCase": True}]
+    }
+    result = pandas_accessor.get_data("test_var", sample_df, payload, MockDataFormat.LIST)
+    filtered_data = pd.DataFrame(result['value']['data'])
+    
+    assert len(filtered_data) == 1
+    assert filtered_data.iloc[0]['StringCol'] == 'Apple'
+
+def test_equals_case_insensitive(pandas_accessor, sample_df):
+    payload = {
+        "filters": [{"col": "StringCol", "value": "apple", "action": "==", "matchCase": False}]
+    }
+    result = pandas_accessor.get_data("test_var", sample_df, payload, MockDataFormat.LIST)
+    filtered_data = pd.DataFrame(result['value']['data'])
+    
+    assert len(filtered_data) == 2
+    assert 'Apple' in filtered_data['StringCol'].values
+    assert 'apple' in filtered_data['StringCol'].values
+
+def test_not_equals_case_insensitive(pandas_accessor, sample_df):
+    payload = {
+        "filters": [{"col": "StringCol", "value": "apple", "action": "!=", "matchCase": False}]
+    }
+    result = pandas_accessor.get_data("test_var", sample_df, payload, MockDataFormat.LIST)
+    filtered_data = pd.DataFrame(result['value']['data'])
+    
+    assert len(filtered_data) == 2
+    assert 'Banana' in filtered_data['StringCol'].values
+    assert 'Cherry' in filtered_data['StringCol'].values
 
 def test_decimator(gui: Gui, helpers, small_dataframe):
     a_decimator = ScatterDecimator(threshold=1)  # noqa: F841
